TGBot/requirements.txt
aiofiles
aiogram
Django
vkbottle
vkbottle-types
aiohttp
aiohappyeyeballs
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
asgiref==3.9.0
attrs==25.3.0
babel==2.17.0
certifi==2025.6.15
choicelib==0.1.5
colorama==0.4.6
django-jazzmin==3.0.1
djangorestframework
frozenlist==1.7.0
idna==3.10
loguru==0.7.3
msgspec==0.19.0
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
vbml==1.1.post1
yarl==1.20.1
openai
httpx_socks
pycountry
deep-translator[google]
redis
aioredis
celery
pytest
pytest-asyncio
